requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.1.0
django-crum==0.7.9
djangorestframework==3.14.0
packaging==23.1
pandas==2.0.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
